Support opening a cmdline window with input history
